git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1020 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/test/tBinaryIO.cpp
@@ -326,7 +326,7 @@ static void testStringSerialization() {
 
         BinaryOutput bo("<memory>", G3D_LITTLE_ENDIAN);
         std::string src = "Hello";
-        bo.writeUInt32(src.length() + 1);
+        bo.writeUInt32((uint32)src.length() + 1);
         bo.writeString(src);
         bo.commit(data);
